update mongo stuff
