+++ composer.json
@@ -19,7 +19,9 @@
   "require": {
     "php": "^5.6.0 || ^7.0",
     "ext-curl": "*",
-    "ext-json": "*",
+    "ext-json": "*"
+  },
+  "suggest": {
     "mongodb/mongodb": "^1.0.0"
   },
   "autoload": {
